Added Buyske algorithm
In rosetta_v2
Added the algorithm that minimizes teh Frobenius norm of the observed matrix and the nearest PSD matrix.
The Buyske algorithm is actived by a new parameter to the rosetta function called missing_corr, which defaults to 'normal' but will use the Buyske routine when missing_corr is set to 'missing' when invoked.
In namespace, added reference to he 'lhs.design' function from the library DoE.wrapper to the NAMESPACE file.